change redis driver, 10k rps

+++ internal/core/domain/hacker.go
@@ -1,6 +1,6 @@
 package domain
 
 type Hacker struct {
-	Name  string `json:"name"`
-	Score int    `json:"score"`
+	Name  string  `json:"name"`
+	Score float64 `json:"score"`
 }
